+++ packages/core/src/constructs.ts
@@ -1366,7 +1366,9 @@ export function or(
       state: DocState<undefined | [number, ParserResult<unknown>]>,
       _defaultValue?,
     ) {
+      let brief: Message | undefined;
       let description: Message | undefined;
+      let footer: Message | undefined;
       let fragments: readonly DocFragment[];
 
       if (state.kind === "unavailable" || state.state == null) {
@@ -1384,7 +1386,9 @@ export function or(
           innerState,
           undefined,
         );
+        brief = docFragments.brief;
         description = docFragments.description;
+        footer = docFragments.footer;
         fragments = docFragments.fragments;
       }
       const entries: DocEntry[] = fragments.filter((f) => f.type === "entry");
@@ -1398,7 +1402,9 @@ export function or(
         }
       }
       return {
+        brief,
         description,
+        footer,
         fragments: [
           ...sections.map<DocFragment>((s) => ({ ...s, type: "section" })),
           { type: "section", entries },
@@ -1892,6 +1898,7 @@ export function longestMatch(
       state: DocState<undefined | [number, ParserResult<unknown>]>,
       _defaultValue?,
     ) {
+      let brief: Message | undefined;
       let description: Message | undefined;
       let footer: Message | undefined;
       let fragments: readonly DocFragment[];
@@ -1907,6 +1914,7 @@ export function longestMatch(
           const docResult = parsers[i].getDocFragments(
             { kind: "available", state: result.next.state },
           );
+          brief = docResult.brief;
           description = docResult.description;
           footer = docResult.footer;
           fragments = docResult.fragments;
@@ -1917,7 +1925,7 @@ export function longestMatch(
         }
       }
 
-      return { description, fragments, footer };
+      return { brief, description, fragments, footer };
     },
   };
 }